+++ tests/ais.rs
@@ -2,6 +2,14 @@
 mod parser;
 #[path = "ais/type1.rs"]
 mod type1;
+#[path = "ais/type11.rs"]
+mod type11;
+#[path = "ais/type12.rs"]
+mod type12;
+#[path = "ais/type13.rs"]
+mod type13;
+#[path = "ais/type15.rs"]
+mod type15;
 #[path = "ais/type18.rs"]
 mod type18;
 #[path = "ais/type19.rs"]
@@ -16,3 +24,11 @@ mod type27;
 mod type4;
 #[path = "ais/type5.rs"]
 mod type5;
+#[path = "ais/type6.rs"]
+mod type6;
+#[path = "ais/type7.rs"]
+mod type7;
+#[path = "ais/type8.rs"]
+mod type8;
+#[path = "ais/type9.rs"]
+mod type9;